move client.js inside src folder
