+++ client/src/index.js
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { ApolloProvider } from "@apollo/client";
-import client from '../client';
+import client from './client';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
